Backend/requirements.txt
Django
django-browser-reload
djangorestframework
Faker
python-decouple
djoser